+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Androidpayments">
         <activity
-            android:name=".AddPaymentCardActivity"
+            android:name=".ui.AddPaymentCardActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Androidpayments">
